new branch

+++ src/main/java/org/wise/portal/service/project/impl/ProjectServiceImpl.java
@@ -131,11 +131,11 @@ public class ProjectServiceImpl implements ProjectService {
     }
   }
 
-  public SharedOwner addSharedTeacher(Long projectId, String username)
+  public SharedOwner addSharedTeacher(Long projectId, String teacherUsername)
       throws ObjectNotFoundException, TeacherAlreadySharedWithProjectException {
+    User user = userDao.retrieveByUsername(teacherUsername);
     Project project = this.getById(projectId);
-    User user = userService.retrieveUserByUsername(username);
-    if (!project.isSharedTeacher(user)) {
+    if (!project.getSharedowners().contains(user)) {
       project.getSharedowners().add(user);
       this.projectDao.save(project);
       this.aclService.addPermission(project, ProjectPermission.VIEW_PROJECT, user);
@@ -144,29 +144,27 @@ public class ProjectServiceImpl implements ProjectService {
       return new SharedOwner(user.getId(), user.getUserDetails().getUsername(),
         user.getUserDetails().getFirstname(), user.getUserDetails().getLastname(), newPermissions);
     } else {
-      throw new TeacherAlreadySharedWithProjectException(user.getUserDetails().getUsername()
-          + " is already shared with this project");
+      throw new TeacherAlreadySharedWithProjectException(teacherUsername + " is already shared with this project");
     }
   }
 
-  public void removeSharedTeacherFromProject(Project project, User user) {
-    removeSharedTeacherAndPermissions(project, user);
-    projectDao.save(project);
-  }
+  public void removeSharedTeacherFromProject(String username, Project project) {
+    User user = userService.retrieveUserByUsername(username);
+    if (project == null || user == null) {
+      return;
+    }
 
-  private void removeSharedTeacherAndPermissions(Project project, User user) {
-    removeSharedTeacher(project, user);
-    removePermissions(project, user);
-  }
-
-  private void removeSharedTeacher(Project project, User user) {
-    project.getSharedowners().remove(user);
-  }
-
-  private void removePermissions(Project project, User user) {
-    List<Permission> permissions = aclService.getPermissions(project, user);
-    for (Permission permission : permissions) {
-      aclService.removePermission(project, permission, user);
+    if (project.getSharedowners().contains(user)) {
+      project.getSharedowners().remove(user);
+      projectDao.save(project);
+      try {
+        List<Permission> permissions = aclService.getPermissions(project, user);
+        for (Permission permission : permissions) {
+          aclService.removePermission(project, permission, user);
+        }
+      } catch (Exception e) {
+        // do nothing. permissions might get be deleted if user requesting the deletion is not the owner of the project.
+      }
     }
   }
 
@@ -274,10 +272,6 @@ public class ProjectServiceImpl implements ProjectService {
 
   public List<Project> getSharedProjectList(User user) {
     return projectDao.getProjectListByUAR(user, "sharedowner");
-  }
-
-  public List<Project> getSharedProjectsWithoutRun(User user) {
-    return projectDao.getSharedProjectsWithoutRun(user);
   }
 
   public String getSharedTeacherRole(Project project, User user) {
@@ -595,8 +589,7 @@ public class ProjectServiceImpl implements ProjectService {
 
   public void removeSharedTeacher(Long projectId, String username)
       throws ObjectNotFoundException {
-    removeSharedTeacherFromProject(getById(projectId),
-        userService.retrieveUserByUsername(username));
+    removeSharedTeacherFromProject(username, getById(projectId));
   }
 
   public void addSharedTeacherPermission(Long projectId, Long userId, Integer permissionId)
@@ -615,36 +608,6 @@ public class ProjectServiceImpl implements ProjectService {
     if (project.getSharedowners().contains(user)) {
       this.aclService.removePermission(project, new ProjectPermission(permissionId), user);
     }
-  }
-
-  @Transactional
-  public void transferProjectOwnership(Project project, User newOwner)
-      throws ObjectNotFoundException {
-    User oldOwner = project.getOwner();
-    if (project.isSharedTeacher(newOwner)) {
-      removeSharedTeacherAndPermissions(project, newOwner);
-    }
-    setOwner(project, newOwner);
-    addSharedTeacherWithViewAndEditPermissions(project, oldOwner);
-    removeAdministrationPermission(project, oldOwner);
-    projectDao.save(project);
-  }
-
-  private void setOwner(Project project, User user) {
-    project.setOwner(user);
-    aclService.addPermission(project, BasePermission.ADMINISTRATION, user);
-  }
-
-  private void addSharedTeacherWithViewAndEditPermissions(Project project, User user) {
-    if (!project.isSharedTeacher(user)) {
-      project.getSharedowners().add(user);
-    }
-    aclService.addPermission(project, ProjectPermission.VIEW_PROJECT, user);
-    aclService.addPermission(project, ProjectPermission.EDIT_PROJECT, user);
-  }
-
-  private void removeAdministrationPermission(Project project, User user) {
-    aclService.removePermission(project, BasePermission.ADMINISTRATION, user);
   }
 
   public List<Project> getProjectsWithoutRuns(User user) {
